Atualizar o 2023_11_14_203632_create_likes_table.php

+++ database/migrations/2023_11_14_203632_create_likes_table.php
@@ -6,15 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
+        if (Schema::hasTable('likes')) {
+            return;
+        }
+
         Schema::create('likes', function (Blueprint $table) {
             $table->id();
+
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('liked_user_id');
+
+            $table->index('user_id');
+            $table->index('liked_user_id');
+
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
@@ -22,9 +28,6 @@ return new class extends Migration
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('likes');
